requirements.txt
altair==4.2.2
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.5
boto3==1.26.66
botocore==1.29.66
branca==0.6.0
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
comm==0.1.2
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.3.0
email-validator==1.3.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
executing==1.2.0
Faker==17.0.0
fastapi==0.92.0
fastjsonschema==2.16.2
favicon==0.7.0
folium==0.14.0
fonttools==4.38.0
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.30
great-expectations==0.15.46
gunicorn==20.1.0
h11==0.14.0
htbuilder==0.6.1
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipykernel==6.21.1
ipython==8.9.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.2.1
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
lxml==4.9.2
makefun==1.15.0
Markdown==3.4.1
markdown-it-py==2.1.0
markdownlit==0.0.5
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.7.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
more-itertools==9.0.0
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
openpyxl==3.1.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==11.0.0
pycparser==2.21
pydantic==1.10.4
pydeck==0.8.0
Pygments==2.14.0
PyJWT==2.6.0
pymdown-extensions==9.9.2
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.1
pytest-html==3.2.0
pytest-metadata==2.0.4
python-dateutil==2.8.2
python-decouple==3.7
python-json-logger==2.0.4
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.3.1
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scipy==1.10.0
semver==2.13.0
Send2Trash==1.8.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
st-annotated-text==3.0.0
stack-data==0.6.2
starlette==0.25.0
streamlit==1.17.0
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.4
streamlit-embedcode==0.1.2
streamlit-extras==0.2.6
streamlit-faker==0.0.2
streamlit-folium==0.11.0
streamlit-image-coordinates==0.1.3
streamlit-keyup==0.2.0
streamlit-option-menu==0.3.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
terminado==0.17.1
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.14
uvicorn==0.20.0
validators==0.20.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.5
zipp==3.12.1
